@@ -5,17 +5,23 @@
     "workspaces/*"
   ],
   "scripts": {
+    "lint": "yarn workspace @yukilabs/governance-frontend lint && yarn workspace @yukilabs/governance-backend lint && yarn workspace @yukilabs/governance-components lint",
     "frontend": "yarn workspace @yukilabs/governance-frontend dev",
     "backend": "yarn workspace @yukilabs/governance-backend dev",
     "components": "yarn workspace @yukilabs/governance-components dev",
     "dev": "concurrently \"yarn frontend\" \"yarn backend\""
   },
   "devDependencies": {
-    "concurrently": "^8.0.1"
+    "concurrently": "^8.1.0"
   },
   "packageManager": "yarn@3.5.1",
   "dependencies": {
-    "typescript": "^5.0.4"
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "concurrently": "^8.0.1",
+    "eslint": "^8.41.0",
+    "eslint-plugin-react": "^7.32.2",
+    "typescript": "^5.1.3"
   },
   "resolutions": {
     "@onflow/fcl@^1.3.2": "patch:@onflow/fcl@npm%3A1.3.2#./.yarn/patches/@onflow-fcl-npm-1.3.2-906125419f.patch"
