@@ -3,6 +3,7 @@
   "private": true,
   "version": "0.0.0",
   "scripts": {
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "test": "jest",
     "dev": "nodemon --exec ts-node --transpile-only src/index.ts",
     "build": "tsc",
@@ -13,19 +14,19 @@
   },
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
-    "@types/jest": "^29.5.1",
+    "@types/jest": "^29.5.2",
     "@types/node": "^20.2.5",
-    "@types/pg": "^8.10.1",
+    "@types/pg": "^8.10.2",
     "drizzle-kit": "^0.18.1",
     "jest": "^29.5.0",
     "nodemon": "^2.0.22",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.0.4"
+    "typescript": "^5.1.3"
   },
   "dependencies": {
-    "@trpc/server": "^10.28.1",
+    "@trpc/server": "^10.29.1",
     "@types/cookie-parser": "^1.4.3",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
@@ -33,13 +34,15 @@
     "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
+    "dotenv": "^16.1.4",
     "dotenv-cli": "^7.2.1",
-    "drizzle-orm": "^0.26.2",
+    "drizzle-orm": "^0.26.5",
+    "drizzle-zod": "^0.4.3",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
     "node-mocks-http": "^1.12.2",
     "pg": "^8.11.0",
+    "superjson": "^1.12.3",
     "trpc": "^0.11.3",
     "zod": "^3.21.4"
   }
