@@ -1,6 +1,6 @@
 import { test, expect } from '@jest/globals';
 import { appRouter } from '.';
-import { users } from '../db/schema';
+import { users } from '../db/schema/users';
 import { eq } from 'drizzle-orm';
 
 import { db } from '../db/db';
