@@ -1,26 +1,30 @@
-import { ButtonProps, Button as ChakraButton } from "@chakra-ui/react";
-import { scrollIntoView } from "../utils/scrollIntoView";
-import React, { forwardRef } from "react";
+import {
+  ButtonProps,
+  Button as ChakraButton,
+} from '@chakra-ui/react';
+import { scrollIntoView } from '../utils/scrollIntoView';
+import React, { forwardRef } from 'react';
 
 type props = {
   variant:
-    | "solid"
-    | "outline"
-    | "outlineLight"
-    | "outlineRounded"
-    | "ghost"
-    | "primaryHero"
-    | "secondaryHero"
-    | "switch"
-    | "filter"
-    | "filterActive"
-    | "category"
-    | "icon";
+    | 'solid'
+    | 'outline'
+    | 'outlineLight'
+    | 'outlineRounded'
+    | 'ghost'
+    | 'primaryHero'
+    | 'secondaryHero'
+    | 'switch'
+    | 'filter'
+    | 'filterActive'
+    | 'category'
+    | 'switcher'
+    | 'icon';
   children: React.ReactNode;
   toId?: string;
   href?: string;
   isExternal?: boolean;
-  target?: ButtonProps["formTarget"];
+  target?: ButtonProps['formTarget'];
 } & ButtonProps;
 
 export const Button = forwardRef<HTMLButtonElement, props>(
@@ -35,7 +39,7 @@ export const Button = forwardRef<HTMLButtonElement, props>(
 
     return (
       <ChakraButton
-        as={href != null ? "a" : undefined}
+        as={href != null ? 'a' : undefined}
         onClick={handleOnClick}
         ref={ref}
         href={href}
@@ -47,4 +51,4 @@ export const Button = forwardRef<HTMLButtonElement, props>(
   }
 );
 
-Button.displayName = "Button";
+Button.displayName = 'Button';
