@@ -1,19 +1,24 @@
-export { AppBar } from './AppBar';
-export { Heading } from './Heading';
-export { Text } from './Text';
-export { Header } from './Header';
-export { Badge } from './Badge';
-export { Button } from './Button';
-export { Connect } from './Connect';
-export { Logo } from './Logo';
-export { ThemeProvider } from './ThemeProvider';
-export { ListRow } from './ListRow';
-export { ListRowContainer } from './ListRow/ListRowContainer';
-export { NavGroup } from './Navigation/NavGroup';
-export { NavItem } from './Navigation/NavItem';
-export { SearchInput } from './SearchInput';
-export { DelegateCard } from './Card';
-export { Tag } from './Tag';
+export { AppBar } from "./AppBar";
+export { Heading } from "./Heading";
+export { Text } from "./Text";
+export { Header } from "./Header";
+export { Badge } from "./Badge";
+export { Button } from "./Button";
+export { Connect } from "./Connect";
+export { Logo } from "./Logo";
+export { PageTitle } from "./PageTitle";
+export { ThemeProvider } from "./ThemeProvider";
+export { NavGroup } from "./Navigation/NavGroup";
+export { NavItem } from "./Navigation/NavItem";
+export { SearchInput } from "./SearchInput";
+export { DelegateCard } from "./Card";
+export { Tag } from "./Tag";
+export { ContentContainer } from "./ContentContainer";
+export { IconButton } from "./IconButton";
+export { Collapse } from "./Collapse";
+
+export { VoteStat } from "./VoteStat";
+
 export {
   HiOutlineCodeBracketSquare,
   HiOutlineLockClosed,
@@ -23,4 +28,8 @@ export {
   HiOutlineQuestionMarkCircle,
   HiOutlineChatBubbleLeftRight,
   HiOutlineChatBubbleLeftEllipsis,
-} from './Icons';
+  MdOutlineHowToVote,
+  MdOutlineAssignment,
+} from "./Icons";
+
+export { QuillEditor } from "./Editor";
