@@ -3,6 +3,8 @@ import { PageContextServer } from "src/renderer/types";
 
 export async function onBeforeRender(pageContext: PageContextServer) {
   return {
-    pageContext: await getDefaultPageContext(pageContext),
+    pageContext: {
+      ...await getDefaultPageContext(pageContext)
+    },
   };
 }
