@@ -6,4 +6,4 @@ declare interface LocaleProps {
   readonly params: LocaleParams;
 }
 
-declare var YOUTUBE_API_KEY: string;
+declare let YOUTUBE_API_KEY: string;
